+++ lib/livingstyleguide/commands/sass.rb
@@ -8,8 +8,11 @@ LivingStyleGuide.command :scss do |arguments, options, scss|
     end
     document.depend_on file
     document.scss << %Q(@import "#{file}";\n)
+  elsif options[:scope] == "global"
+    document.scss << scss
   else
-    document.scss << "##{document.id.gsub(/[\/\.]/, '\\\\\0')} {\n#{scss}\n}\n"
+    id = document.id.gsub(/[\/\.]/, '\\\\\0')
+    document.scss << "##{id} {\n#{scss}\n}\n"
   end
   nil
 end
@@ -17,14 +20,10 @@ end
 LivingStyleGuide.command :sass do |arguments, options, sass|
   file = arguments.first
   if file
-    if document.file
-      file = File.join(File.dirname(document.file), file)
-    end
-    document.depend_on file
-    document.scss << %Q(@import "#{file}";\n)
+    scss(arguments, options, nil)
   else
-    scss = Sass::Engine.new(sass).to_tree.to_scss
-    document.scss << "##{document.id.gsub(/[\/\.]/, '\\\\\0')} {\n#{scss}\n}\n"
+    converted = Sass::Engine.new(sass).to_tree.to_scss
+    scss(arguments, options, converted)
   end
   nil
 end