+++ test/unit/commands/sass_test.rb
@@ -33,6 +33,21 @@ class CssTest < DocumentTestCase
     CSS
   end
 
+  def test_adding_global_scss
+    doc = LivingStyleGuide::Document.new do <<-INPUT.unindent
+        @scss scope: global {
+          #adding-global-scss {
+            background: red;
+            #x { color: green; }
+          }
+        }
+      INPUT
+    end
+    doc.render
+    assert_match(/(?<!##{doc.id} )#adding-global-scss \{/, doc.css)
+    assert_match(/(?<!##{doc.id} )#adding-global-scss #x \{/, doc.css)
+  end
+
   def test_adding_sass
     doc = LivingStyleGuide::Document.new do <<-INPUT.unindent
         @sass preprocessor: sass
@@ -47,6 +62,20 @@ class CssTest < DocumentTestCase
       ##{doc.id} #adding-sass \{ background: red; \}
       ##{doc.id} #adding-sass #x \{ color: green; \}
     CSS
+  end
+
+  def test_adding_global_sass
+    doc = LivingStyleGuide::Document.new do <<-INPUT.unindent
+        @sass scope: global
+          #adding-global-sass
+            background: blue
+            #x
+              background: yellow
+      INPUT
+    end
+    doc.render
+    assert_match(/(?<!##{doc.id} )#adding-global-sass \{/, doc.css)
+    assert_match(/(?<!##{doc.id} )#adding-global-sass #x \{/, doc.css)
   end
 
   def test_adding_css_inside_examples