+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+* CSS/Sass code can be added to the global scope from now on
+
+
 ## 2.0.0.pre.2
 
 * Color swatches accept nested functions like `lighten($my-color)` from now on